backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
opencv-python>=4.8.1.78
Pillow>=10.1.0
torch>=2.6.0
torchvision>=0.19.0
timm>=0.9.12
transformers>=4.36.2
huggingface-hub>=0.19.4
pandas>=2.1.4
numpy>=1.24.4
scikit-learn>=1.3.2
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
